requirements.txt
python-telegram-bot==6.1.0
pymongo==3.4.0
six>=1.10.0
dateutils==0.6.6
Pillow>=2.0.0
mongoengine>=0.1